+++ NEWS.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+New functionality:
+
+* Added `BigwigTrack()` function to plot data from bigWig files
+* Added `bigwig` and `bigwig.type` arguments to `CoveragePlot()` to
+include bigWig files in `CoveragePlot()`
+
 # Signac 1.1.1 
 
 New functionality: